+++ aios/memory/base.py
@@ -1,96 +1,25 @@
 # This file provides a wrapper on memory access, similarly to working with
 # pointers in low level languages
 # The memory is organized in blocks of a single byte
-from Cerebrum.cerebrum.memory.apis import MemoryQuery
-from aios.syscall.memory import MemorySyscall
+from cerebrum.memory.apis import MemoryQuery
+# Remove the circular import
+# from aios.syscall.memory import MemorySyscall
 from .retrievers import SimpleEmbeddingRetriever, ChromaRetriever
 
 # use C compatible data types for maximum memory efficiency
 import ctypes
-from typing import Dict, Any, Optional
-
-
-class MemoryNote:
-    """A memory note that represents a single unit of information in the memory system.
-    
-    This class encapsulates all metadata associated with a memory, including:
-    - Core content and identifiers
-    - Temporal information (creation and access times)
-    - Semantic metadata (keywords, context, tags)
-    - Relationship data (links to other memories)
-    - Usage statistics (retrieval count)
-    - Evolution tracking (history of changes)
-    """
-    
-    def __init__(self, 
-                 content: str,
-                 id: Optional[str] = None,
-                 keywords: Optional[List[str]] = None,
-                 links: Optional[Dict] = None,
-                 retrieval_count: Optional[int] = None,
-                 timestamp: Optional[str] = None,
-                 last_accessed: Optional[str] = None,
-                 context: Optional[str] = None,
-                 evolution_history: Optional[List] = None,
-                 category: Optional[str] = None,
-                 tags: Optional[List[str]] = None):
-        """Initialize a new memory note with its associated metadata.
-        
-        Args:
-            content (str): The main text content of the memory
-            id (Optional[str]): Unique identifier for the memory. If None, a UUID will be generated
-            keywords (Optional[List[str]]): Key terms extracted from the content
-            links (Optional[Dict]): References to related memories
-            retrieval_count (Optional[int]): Number of times this memory has been accessed
-            timestamp (Optional[str]): Creation time in format YYYYMMDDHHMM
-            last_accessed (Optional[str]): Last access time in format YYYYMMDDHHMM
-            context (Optional[str]): The broader context or domain of the memory
-            evolution_history (Optional[List]): Record of how the memory has evolved
-            category (Optional[str]): Classification category
-            tags (Optional[List[str]]): Additional classification tags
-        """
-        # Core content and ID
-        self.content = content
-        self.id = id or str(uuid.uuid4())
-        
-        # Semantic metadata
-        self.keywords = keywords or []
-        self.links = links or []
-        self.context = context or "General"
-        self.category = category or "Uncategorized"
-        self.tags = tags or []
-        
-        # Temporal information
-        current_time = datetime.now().strftime("%Y%m%d%H%M")
-        self.timestamp = timestamp or current_time
-        self.last_accessed = last_accessed or current_time
-        
-        # Usage and evolution data
-        self.retrieval_count = retrieval_count or 0
-        self.evolution_history = evolution_history or []
-
-    def return_params(self) -> Dict[str, Any]:
-        return {
-            "content": self.content or "",
-            "id": self.id or "",
-            "keywords": self.keywords or [],
-            "links": self.links or [],
-            "retrieval_count": self.retrieval_count or 0,
-            "timestamp": self.timestamp or "",
-            "last_accessed": self.last_accessed or "",
-            "context": self.context or "",
-            "evolution_history": self.evolution_history or [],
-            "category": self.category or "",
-            "tags": self.tags or []
-        }
+from typing import Dict, Any, Optional, List
+import uuid
+from datetime import datetime
 
 # abstract implementation of memory utilities for thread safe access
 class BaseMemoryManager:
-    def __init__(self):
+    def __init__(self,log_mode):
         self.chroma_retriever = ChromaRetriever()
         self.memories = {}
 
-    def _analyze_query_to_memory(self, query: MemoryQuery) -> MemoryNote:
+    def _analyze_query_to_memory(self, query: MemoryQuery) -> 'MemoryNote':
+        from .note import MemoryNote  # Import here to avoid circular dependency
         params = query.params
         valid_keys = ["content", "id", "keywords", "links", "retrieval_count", 
              "timestamp", "last_accessed", "context", "evolution_history", 
@@ -99,7 +28,12 @@ class BaseMemoryManager:
         memory_note = MemoryNote(**filtered_data)
         return memory_note
 
-    def address_request(self, memory_syscall: MemorySyscall):
+    def address_request(self, memory_syscall):
+        # Import here to avoid circular dependency
+        from aios.syscall.memory import MemorySyscall
+        if not isinstance(memory_syscall, MemorySyscall):
+            raise TypeError(f"Expected MemorySyscall, got {type(memory_syscall)}")
+            
         memory_note = self._analyze_query_to_memory(memory_syscall.query)
         if memory_syscall.query.operation == "add_memory":
             self.add_memory(memory_note)
@@ -114,7 +48,11 @@ class BaseMemoryManager:
         else:
             raise ValueError(f"Invalid operation: {memory_syscall.query.operation}")
 
-    def add_memory(self, memory_note: MemoryNote):
+    def add_memory(self, memory_note):
+        from .note import MemoryNote  # Import here to avoid circular dependency
+        if not isinstance(memory_note, MemoryNote):
+            raise TypeError(f"Expected MemoryNote, got {type(memory_note)}")
+            
         metadata = {
             "context": memory_note.context,
             "keywords": memory_note.keywords,
@@ -125,7 +63,11 @@ class BaseMemoryManager:
         self.chroma_retriever.add_document(document=memory_note.content, metadata=metadata, doc_id=memory_note.id)
         return memory_note.id
 
-    def remove_memory(self, memory_note: MemoryNote):
+    def remove_memory(self, memory_note):
+        from .note import MemoryNote  # Import here to avoid circular dependency
+        if not isinstance(memory_note, MemoryNote):
+            raise TypeError(f"Expected MemoryNote, got {type(memory_note)}")
+            
         memory_id = memory_note.id
         if memory_id in self.memories:
             # Delete from ChromaDB
@@ -135,7 +77,11 @@ class BaseMemoryManager:
             return True
         return False
 
-    def update_memory(self, memory_note: MemoryNote):
+    def update_memory(self, memory_note):
+        from .note import MemoryNote  # Import here to avoid circular dependency
+        if not isinstance(memory_note, MemoryNote):
+            raise TypeError(f"Expected MemoryNote, got {type(memory_note)}")
+            
         memory_id = memory_note.id
         if memory_id not in self.memories:
             return False
@@ -154,7 +100,11 @@ class BaseMemoryManager:
         
         return True
 
-    def get_memory(self, memory_note: MemoryNote):
+    def get_memory(self, memory_note):
+        from .note import MemoryNote  # Import here to avoid circular dependency
+        if not isinstance(memory_note, MemoryNote):
+            raise TypeError(f"Expected MemoryNote, got {type(memory_note)}")
+            
         memory_id = memory_note.id
         if memory_id not in self.memories:
             return None
@@ -173,5 +123,3 @@ class BaseMemoryManager:
             retrieved_memories.append(memory)
                 
         return retrieved_memories[:k]
-
-
