+++ aios/memory/note.py
@@ -1,0 +1,82 @@
+"""
+Memory note module containing the MemoryNote class.
+This file is separated from base.py to avoid circular imports.
+"""
+from typing import Dict, Any, Optional, List
+import uuid
+from datetime import datetime
+
+
+class MemoryNote:
+    """A memory note that represents a single unit of information in the memory system.
+    
+    This class encapsulates all metadata associated with a memory, including:
+    - Core content and identifiers
+    - Temporal information (creation and access times)
+    - Semantic metadata (keywords, context, tags)
+    - Relationship data (links to other memories)
+    - Usage statistics (retrieval count)
+    - Evolution tracking (history of changes)
+    """
+    
+    def __init__(self, 
+                 content: str,
+                 id: Optional[str] = None,
+                 keywords: Optional[List[str]] = None,
+                 links: Optional[Dict] = None,
+                 retrieval_count: Optional[int] = None,
+                 timestamp: Optional[str] = None,
+                 last_accessed: Optional[str] = None,
+                 context: Optional[str] = None,
+                 evolution_history: Optional[List] = None,
+                 category: Optional[str] = None,
+                 tags: Optional[List[str]] = None):
+        """Initialize a new memory note with its associated metadata.
+        
+        Args:
+            content (str): The main text content of the memory
+            id (Optional[str]): Unique identifier for the memory. If None, a UUID will be generated
+            keywords (Optional[List[str]]): Key terms extracted from the content
+            links (Optional[Dict]): References to related memories
+            retrieval_count (Optional[int]): Number of times this memory has been accessed
+            timestamp (Optional[str]): Creation time in format YYYYMMDDHHMM
+            last_accessed (Optional[str]): Last access time in format YYYYMMDDHHMM
+            context (Optional[str]): The broader context or domain of the memory
+            evolution_history (Optional[List]): Record of how the memory has evolved
+            category (Optional[str]): Classification category
+            tags (Optional[List[str]]): Additional classification tags
+        """
+        # Core content and ID
+        self.content = content
+        self.id = id or str(uuid.uuid4())
+        
+        # Semantic metadata
+        self.keywords = keywords or []
+        self.links = links or []
+        self.context = context or "General"
+        self.category = category or "Uncategorized"
+        self.tags = tags or []
+        
+        # Temporal information
+        current_time = datetime.now().strftime("%Y%m%d%H%M")
+        self.timestamp = timestamp or current_time
+        self.last_accessed = last_accessed or current_time
+        
+        # Usage and evolution data
+        self.retrieval_count = retrieval_count or 0
+        self.evolution_history = evolution_history or []
+
+    def return_params(self) -> Dict[str, Any]:
+        return {
+            "content": self.content or "",
+            "id": self.id or "",
+            "keywords": self.keywords or [],
+            "links": self.links or [],
+            "retrieval_count": self.retrieval_count or 0,
+            "timestamp": self.timestamp or "",
+            "last_accessed": self.last_accessed or "",
+            "context": self.context or "",
+            "evolution_history": self.evolution_history or [],
+            "category": self.category or "",
+            "tags": self.tags or []
+        }
